requirements.txt
Django>=4.2.20
djangorestframework>=3.14.0
requests>=2.31.0
coreapi>=2.3.3
PyYAML>=6.0
